added option to change angle and distance of GLCM

+++ examples/classes.py
@@ -3,7 +3,7 @@ if __name__ == "__main__":
 
     whole=SlidingWindow("./input.tif",13,"homogeneity",downscale=1)
     whole.render("tiling",tilesz=125)
-    whole.savetif("outputclasstiling", type="uint8")
+    whole.savetif("outputclasstiling", type="float32")
 
     downscale=SlidingWindow("./input.tif",13,"homogeneity",downscale=2)
     downscale.render("raster",rowssave=80)
